requirements.txt
torch
torchvision
tqdm
Pillow
scikit-learn
matplotlib
gradio